Crate pallet_evm_tracker

Pallet EVM Account nonce tracker

Re-exports

pub use check_nonce::CheckNonce;

Modules

check_nonce

create_contract

Contract creation allow list implementations

dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

fees

Fees module for EVM domain

storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

traits

Traits that make programming against the runtime easier.

Structs

DefaultToAnyone

Default value for ContractCreationAllowedBy if it is not set.

Pallet

Pallet EVM account nonce tracker.

Enums

Call

Contains a variant per dispatchable extrinsic that this pallet has.

Traits

Config

Configuration trait of this pallet.

Type Aliases

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.