pallet_evm_tracker

Module create_contract

Contract creation allow list implementations

Structs

CheckContractCreation

Reject contract creation, unless the account is in the current evm contract allow list.

Enums

Pre

Data passed from prepare to post_dispatch.

Val

Data passed from validate to prepare.

Functions

is_create_contract

Returns true if the call is a contract creation call.

is_create_contract_allowed

Rejects contracts that can’t be created under the current allow list. Returns false if the call is a contract call, and the account is not allowed to call it. Otherwise, returns true.

is_create_unsigned_contract_allowed

If anyone is allowed to create contracts, allows contracts. Otherwise, rejects contracts. Returns false if the call is a contract call, and there is a specific (possibly empty) allow list. Otherwise, returns true.